Time: 0 ms (100.00%), Space: 6.9 MB (97.07%) - LeetHub

+++ 1491-average-salary-excluding-the-minimum-and-maximum-salary/1491-average-salary-excluding-the-minimum-and-maximum-salary.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     double average(vector<int>& salary) {
         
-        double min = INT_MAX;
-        double max = INT_MIN; 
+        int min = INT_MAX;
+        int max = INT_MIN; 
         double sum = 0;
         int i;
         for(i= 0; i < salary.size();i++)
